+++ src/main/java/com/t1membership/item/dto/registerItem/RegisterItemRes.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -14,7 +16,7 @@ public class RegisterItemRes {
 
     private String itemName;
 
-    private int itemPrice;
+    private BigDecimal itemPrice;
 
     private int itemStock;
